fix: error message before choosing a provider
when user logs in without having a stored selected provider, they should be directed to a page where to select a provider
