+++ src/scenes/App/useNoSelectedProvider.ts
@@ -1,0 +1,9 @@
+import { useAppSelector } from '../../store/hooks';
+
+export const useNoSelectedProvider = () => {
+  const provider = useAppSelector(
+    (state) => state.userContext.activeProviderCode,
+  );
+
+  return !provider;
+};
